requirements.txt
Django==1.8.1
Pillow==2.8.1
django-debug-toolbar==1.3.0
